Added some additional states along with sprite hb
Added some more state scripts in for the enemy object and tweaked a few things for the player scripts. I also added a temporary hitbox for the player attack sprite and now have it to where when the player attacks the enemy if the enemy is in the players attack hitbox, the enemy dies.