Reworked player movement
Reworked player movement back to the step event in the object for now, I also reworked how colliding with the environment works in an attempt to make it so the collision code can be used for player and enemy objects.

Still plan on working on enemy following player, but needed to do this rework first.