Fixed issue of enemy movement
Resolved the issue I was having that was causing multiple enemies to move at the same intervals when changing move_speed and jump_speed. Now I just have an issue of fine tuning the ai so that the enemy objects do not go over edges even if they would be "chasing" the player.